requirements.txt
flask
google-generativeai>=0.8.0
gunicorn